Rework Buildkite bootstrapping
